requirements.txt
azure-identity
azure-storage-file-datalake
boto3
requests
